store matrix, and column headers for the page.  Store the start time of the page.
Move all the time calculations to the server side.  This will help with consistency of timekeeping.   When the JS was taking the time there is no guarantee that the clocks of the client and server are perfectly in sync.

Storing start time and click counters on the player object as well.  Now the table page is robust to reloads.  That is, the numbers and column headers and timing information remain intact when the page is reloaded.